feat(at_auth): remove at_keys.dart from at_commons and place in at_auth.dart
