+++ packages/at_auth/lib/src/auth/models/at_auth_responses.dart
@@ -1,4 +1,4 @@
-import 'package:at_auth/src/keys/at_keys.dart';
+import 'package:at_auth/src/keys/at_keys_legacy.dart';
 import 'package:at_chops/at_chops.dart';
 import 'package:at_lookup/at_lookup.dart';
