+++ packages/at_auth/lib/src/keys/at_keys_io_impl.dart
@@ -4,7 +4,7 @@ import 'dart:io';
 
 import 'package:at_chops/at_chops.dart';
 import 'package:at_commons/at_commons.dart';
-import 'package:at_auth/src/keys/at_keys.dart';
+import 'package:at_auth/src/keys/at_keys_legacy.dart';
 import 'package:at_auth/src/keys/at_keys_io.dart';
 import 'package:at_auth/src/exception/at_auth_exceptions.dart';
