+++ packages/at_auth/test/enrollment_test.dart
@@ -1,7 +1,11 @@
 import 'dart:convert';
 
-import 'package:at_auth/at_auth.dart';
+import 'package:at_auth/src/enroll/at_enrollment.dart';
 import 'package:at_auth/src/enroll/at_enrollment_impl.dart';
+import 'package:at_auth/src/enroll/models/at_enrollment_request.dart';
+import 'package:at_auth/src/enroll/models/at_enrollment_response.dart';
+import 'package:at_auth/src/enroll/models/enrollment_request_decision.dart';
+import 'package:at_auth/src/keys/at_keys_legacy.dart';
 import 'package:at_chops/at_chops.dart';
 import 'package:at_commons/at_builders.dart';
 import 'package:at_commons/at_commons.dart';
